requirements.txt
fastapi
uvicorn[standard]
pdfplumber
pytesseract
Pillow
spacy
streamlit
python-multipart
